Avoid adding dynamic listeners to writer
When the writer loses all connections, while it waits for a
reconnection, each message it receives to publish creates a
new ready listener. This can cause node to warn about a
possible listener leak if too many messages are received.
To avoid that, we instead add the messages to an array that
will be drained after the ready event is sent.